Create Knight#move_finder
This method finds the minimum number of moves needed to get from the provided location to the provided destination.

It first checks if it is already at the destination, then it checks if the destination is possible from the current location. If neither of this is possible, it resorts to recursion to find the required number of moves using breath first traversal.